Campaign small fixes
Added restrictions for when the player can and cannot use docks or tradelanes/certain docks or tradelanes. Autopilot still tries to do the path but nothing will happen, tho this will be fixed when the autopilot is refactor.

Gave EmbeddedServer.cs the saveFolder varible so that SaveSP and Save and create save games on the proper save folder without the need of searching for the path again.

Improved code for mission check on load or start game to avoid crashes on invalid mission numbers, refactored WriteActiveTriggers to write all triggers on base save or write only the active one in space.